Reorganize codebase and enhance build process
This commit modifies the file paths in the index.d.ts file, relocating certain exports to a new "src" directory. Additionally, the build process in the package.json file is updated, introducing a 'tsc' command for compiling typescript files and generating a new target "dist" directory. To ensure consistency, all associated file paths in the renamed directories have been updated accordingly.
