+++ index.d.ts
@@ -1,10 +1,10 @@
 // Version: 1.0.0
 
 export = DApiGate;
-export * from "./rate";
-export * from "./proxy";
-export * from "./phone";
-export * from "./api";
+export * from "./src/rate";
+export * from "./src/proxy";
+export * from "./src/phone";
+export * from "./src/api";
 
 export as namespace DApiGate;
