+++ src/rate/index.d.ts
